Rename BasicEntityCheckboxField from temp name
