+++ src/lib/BasicEntityField.tsx
@@ -1,11 +1,9 @@
-import React, { useEffect, forwardRef } from "react";
-
-import { TextField } from "@mui/material";
+import React from "react";
 
 import type { Spec } from "./basic-types";
 
 import { Default, Exact, Gubu, Skip } from "gubu";
-import { BasicEntityCheckboxField } from "./BasicEntityCheckField";
+import { BasicEntityCheckboxField } from "./BasicEntityCheckboxField";
 import { BasicEntityAutocompleteField } from "./BasicEntityAutocompleteField";
 import { BasicEntitySliderField } from "./BasicEntitySliderField";
 import { BasicEntityRadioGroupField } from "./BasicEntityRadioGroupField";
